Code restructure to make methods more maintainable, split project, compound, and json methods into their own files.
