+++ Chemlist/Chemlist.csproj
@@ -65,6 +65,15 @@
     <Compile Include="Compounds\AddNewCompound.Designer.cs">
       <DependentUpon>AddNewCompound.cs</DependentUpon>
     </Compile>
+    <Compile Include="ObjectsAndMethods\CompoundMethods.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ObjectsAndMethods\JsonHandler.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ObjectsAndMethods\ProjectMethods.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Projects\AddNewProject.cs">
       <SubType>Form</SubType>
     </Compile>
